some changes on popup and filterable list

+++ js/modules/Popup.js
@@ -60,6 +60,8 @@ export default class Popup{
                             <p class="modal-inputDescription uk-margin-remove-bottom">Bitte beschreibe den Schaden kurz!</p>
                             <textarea class="uk-textarea uk-height-small"></textarea>
                         </div>
+                        <div class="report-errorMessage"></span></div>
+
                     `;
                     returnButtonTxt = "Abbrechen";
                     proceedButtonTxt = "Speichern";
@@ -69,7 +71,8 @@ export default class Popup{
                     titleIcon = `<svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 20 20">
                                     <path id="Pfad_23" data-name="Pfad 23" d="M2.679,39.363a.257.257,0,0,0,.439,0L5.734,35.8c.121-.165.064-.3-.128-.3H4.218a.3.3,0,0,1-.281-.379,7.487,7.487,0,0,1,7.079-6.337c4,0,7.249,3.619,7.249,8.068s-3.252,8.068-7.249,8.068a.971.971,0,0,0,0,1.932c4.954,0,8.985-4.486,8.985-10s-4.03-10-8.985-10c-4.423,0-8.107,3.576-8.847,8.266a.439.439,0,0,1-.4.382H.192c-.192,0-.249.134-.128.3Z" transform="translate(0 -26.855)" fill="#07f"/>
                                 </svg>`;
-                    content = `<input type="text" readonly="readonly" id="datepicker-popup" placeholder="Tag auswählen">`;
+                    content = `<input type="text" readonly="readonly" id="datepicker-popup" placehol
+der="Tag auswählen">`;
                     returnButtonTxt = "Abbrechen";
                     proceedButtonTxt = "Verlängern";
                     break;
@@ -123,12 +126,9 @@ export default class Popup{
                     this.vars.equipment = equipment;
                     const reservation = await this.modules.ajax.getResById(this.vars.resId);
                     this.vars.reservation = reservation;
-                    this.vars.originalResStatus = reservation.statusId;
 
                     if(!equipment || !reservation){
                         this.vars.equipmentRequest = false;
-                        this.vars.reportDamage = "Problem bei der API Abfrage";
-                        this.vars.reportTodo = "Problem bei der API Abfrage";
                         $(".proceedButton").remove();
                     }else{
                         this.vars.equipmentRequest = true;
@@ -138,12 +138,14 @@ export default class Popup{
                 
                     //for first state
                     this.vars.reportState = "damage";
+                    console.log(this.vars.equipmentRequest)
+                    console.log(this.vars.equipment);
                     if(this.vars.equipmentRequest){
                         //if equipment request was sucessfull insert value
+                        textArea.attr("placeholder", "Schadensmeldung");
                         textArea.val(this.vars.reportDamage);
                     }else{
-                        //if not insert the value as placeholder beause it contains error message
-                        textArea.attr("placeholder", this.vars.reportDamage);
+                        textArea.attr("placeholder", "Fehler bei der API Abfrage!");
                     }
                     
                     //on change of radio button changes state of report popup and adds funcions
@@ -155,21 +157,25 @@ export default class Popup{
                         if(val == "damage"){
                             //set value to textarea
                             $(".modal-inputDescription").text("Bitte beschreibe den Schaden kurz!");
+                            textArea.attr("placeholder", "Schadensmeldung");
                             if(this.vars.equipmentRequest){
                                 textArea.val(this.vars.reportDamage);
                             }else{
-                                textArea.attr("placeholder", this.vars.reportDamage);
+                                console.log(this.vars.reportDamage);
+                                textArea.attr("placeholder", "Fehler bei der API Abrage!");
                             }
                         }else if(val == "todo"){
                             //set value to textarea
                             $(".modal-inputDescription").text("Beschreibe kurz, was zu tun ist.");
+                            textArea.attr("placeholder", "Todo Eintrag");
                             if(this.vars.equipmentRequest){
+                                this.vars.reportTodo ? textArea.val(this.vars.reportTodo) :
                                 textArea.val(this.vars.reportTodo);
                             }else{
-                                textArea.attr("placeholder", this.vars.reportTodo);
+                                textArea.attr("placeholder", "Fehler bei der API Abfrage");
                             }
                         }else if(val == "cancel"){
-                            $(".modal-inputDescription").text("Der Status dieser Reservierung wird auf gelöscht gesetzt!");
+                            $(".modal-inputDescription").text("Diese Reservierung wird storniert. Achtung, kann nicht rückgängig gemacht werden!");
                             textArea.val("");
                             textArea.attr("placeholder", "Stornierungsgrund");
                         }
@@ -209,39 +215,53 @@ export default class Popup{
         async addProceedFunction(){
             //if proceed button is clicked
             $(".proceedButton").on("click", async () => {
-                let success = false;
+                const errorText = $(".report-errorMessage");
+                let empty = true;
+                let apiAnswer = false;
                 switch (this.vars.popupType){
                     case "report":
                         if(this.vars.reportState == "damage"){
-                            this.vars.equipment.damage = this.vars.reportDamage;
-                            const equipPut = await this.modules.ajax.putEquipment(this.vars.eqId, this.vars.equipment);
-                            if(equipPut)success = true;
+                            if(this.vars.reportDamage != "" && this.vars.reportTodo != undefined){
+                                empty = false;
+                                this.vars.equipment.damage = this.vars.reportDamage;
+                                apiAnswer = await this.modules.ajax.putEquipment(this.vars.eqId, this.vars.equipment);
+                                console.log(apiAnswer);
+                            }
                         }else if(this.vars.reportState == "todo"){
-                            this.vars.equipment.todo = this.vars.reportTodo;
-                            const apiAnswer = await this.modules.ajax.putEquipment(this.vars.eqId, this.vars.equipment);
-                            if(apiAnswer)success = true;
-                            console.log(apiAnswer)
+                            if(this.vars.reportTodo != "" && this.vars.reportTodo != undefined){
+                                empty = false;
+                                this.vars.equipment.todo = this.vars.reportTodo;
+                                apiAnswer = await this.modules.ajax.putEquipment(this.vars.eqId, this.vars.equipment);
+                                console.log(apiAnswer)
+                            }
                         }else if(this.vars.reportState == "cancel"){
-                            const cancelValue = [{
-                              "id": this.vars.resId*1,
-                              "grund": this.vars.reportCancel
-                            }];
-                            const cancelRequest = await this.modules.ajax.deleteReservation(cancelValue);
-                            console.log(cancelRequest)
-                            if(cancelRequest)success = true;
+                            if(this.vars.reportCancel != "" && this.vars.reportCancel != undefined){
+                                empty = false;
+                                const cancelMessage = `<h2 class="uk-text-large">Wirklich stornieren?</h2><p>Bist du sicher, dass du diese Reservierung stornieren möchtest? Das kann später nicht mehr rückgängig gemacht werden!</p>`
+                                UIkit.modal.confirm(cancelMessage, {stack:true}).then(async () => {
+                                    const cancelValue = [{
+                                        "id": this.vars.resId*1,
+                                        "grund": this.vars.reportCancel
+                                    }];
+                                    apiAnswer = await this.modules.ajax.deleteReservation(cancelValue);
+                                    console.log(apiAnswer);
+                                }, function () {
+                                    console.log('Rejected.');
+                                });
+                            }
                         }
                         break;
                     case "extend":
                         if(this.vars.selectedDate){
-                            const extendAnswer = this.modules.ajax.extendReservation(this.vars.resId, this.vars.selectedDate);
-                            if(!extendAnswer)success = false;
+                            apiAnswer = this.modules.ajax.extendReservation(this.vars.resId, this.vars.selectedDate);
                         }else{
                           $("#datepicker-popup").val("Kein gültiges Datum!");
                           const d = new Date(this.vars.selectedDate);
                         }
                         break;
                 }
-                if(success){
+                console.log(apiAnswer);
+                if(apiAnswer){
                     UIkit.modal(`#${this.vars.modalId}`).hide();
                     this.removePopup();
                     console.log(this.vars.reportState);
@@ -254,7 +274,11 @@ export default class Popup{
                         this.displayUndoCancel(resContainer, this.vars.resId);
                     }
                 }else{
-                    //this.doms.proceedButton.text("Fehlgeschlagen");
+                    if(empty){
+                        errorText.html("<span uk-icon=\"icon: warning\"></span><span>Das Textfeld darf nicht leer bleiben!</span>");
+                    }else{
+                        errorText.html("<span uk-icon=\"icon: warning\"></span><span>Leider konnte die Änderung gerade nicht durchgeführt werden!</span>");
+                    }
                 }
             });
         }
